Final Project: Formatting the age string correctly
