+++ patients/templates/patients/profile.html
@@ -3,7 +3,7 @@
 {% block title %} {{ patient.user.first_name }} {{ patient.user.last_name }} {% endblock %}
 
 {% block body %}
-    <h1>{{ patient.user.first_name }}</h1>
+    <h1>{{ patient.user.first_name }} {{ patient.user.first_name }}</h1>
     <table>
         <tr>
             <th>Stat</th>
