+++ patients/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render
+from django.http import HttpResponse
 from django.contrib.auth.decorators import login_required 
-from users.models import Patient
+from patients.models import Patient
 
 # Create your views here.
